Route to single event, improvement design
route:
in pages a events folder with index.vue as the page with all the events
and with _event.vue as the page with a single event.
NuxtLink routes to a single event with a parameter, in the single event
the fetch gets the clicked event based on id parameter.

design:
rainbow